Added all of the necessary math functions and prompts.
Problems to consider: input needs to have .strip() applied so that they
dont misbehave. Additionally, it would be useful to reconfigure the
structure of the program so that it functions more as a calculator which
runs until commanded otherwise. Also, the prompts for input should be
moved around so that it makes a bit more sense.